move hook out to separate file
