+++ packages/scenes-react/src/index.ts
@@ -6,4 +6,5 @@ export { VizPanel } from './components/VizPanel';
 export { RefreshPicker } from './components/RefreshPicker';
 export { CustomVariable } from './variables/CustomVariable';
 export { Breadcrumb, BreadcrumbContext, BreadcrumbProvider } from './contexts/BreadcrumbContext';
+export { useVariableValues } from './hooks/useVariableValues';
 export * from './hooks/hooks';
